fix: improve some styles on users page
